+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## 2022.08.18
-* Add go client
+* Create client
+* Add gateway, firewall and nat rules management